requirement.txt
cryptography
Django
django-allauth
django-cache
django-extensions
django-filter
django-js-asset
django-mptt
gunicorn
idna
idpay
Jinja2
kombu
lxml
Pillow
psycopg2-binary
python3-openid
pytz
redis
requests
social-auth-app-django
social-auth-core
urllib3
user-agents
python-decouple
django-sentry
sentry-sdk
django-user-agents
django-debug-toolbar